battleshit done
could add more for battle shit but fine for now, wanted to make more of a kitty litter box, if i can get around to it. but right now priority is on making some more mini games,

type cat where is it just meow meow meow? would be fun, or actual words and it comes out as meow meow meow